requirements.txt
crewai
langchain_openai
crewai_tools
python-dotenv
streamlit
 